Bug 8945 - Missing key is getVrfId. Supplied key is VpnInstanceOpDataEntryKey
On a router delete, the internal VPN for that router was being removed
before handling external networks associated to that router.
Releasing router based lport tags, which is part of L3VPN Over VxLAN,
depends on vpn operational data, but the vpn data was being deleted much
before handling external networks for that router.

Change-Id: I72c0ff142b3929808afcfa7e49589b5420f48510
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>