BUG 8969 - Fix Exeption when clearing external router GW
When clearing the router GW, the flow entry for router GW
on the NAPT primary switch is also removed.
Exception occurs when we try to delete the router GW
but there is no NAPT primary switch.

Fix: check for existence of NAPT primary switch
before deleting the router GW flow on the switch.
Also downgrade ERROR to DEBUG log level when primary
NAPT switch can not be found since this could
happen when no internal network attaches to the router yet.

Change-Id: I7ebc0d60183954044aeb8a4e8f062b4d227f6c76
Signed-off-by: Vinh Nguyen <vinh.nguyen@hcl.com>