neutronvpn: create ipv4 or ipv6 context
A VPN instance has ipv4iFamily and ipv6Family attributes. According to
the nature of router with which a VPN was associated, the ipv4 attribute
is instanciated or the ipv6 attribute is instanciated, or both
attributes are instanciated. On VPN instance update, parse the ipv4 or
ipv6 family attributes of the VPN instance and update the RDs
accordingly.

Change-Id: If5ea1a0a29d31388d7898b9e6e53bdc4aecb1cbe
Signed-off-by: Valentina Krasnobaeva <valentina.krasnobaeva@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Noel de Prandieres <prandieres@6wind.com>