neutronvpn: create ipv4 or ipv6 context
A VPN instance node has ipv4 and ipv6 ready attributes. The modification
does the following: according to the nature of the router with which a
VPN is attached, the ipv4 attribute is instanciated or the ipv6
attribute is instanciated, or both attributes are instanciated. On VPN
instance update, parse the ipv4 or ipv6 family attributes of the VPN
instance and update the RDs accordingly.

Change-Id: I2328336a85187f1a61b1ecdb0b812f56cb1996f5
Signed-off-by: Valentina Krasnobaeva <valentina.krasnobaeva@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Noel de Prandieres <prandieres@6wind.com>