Bug8861 : Data validation failed for path /snatint-ip-port-map
Description : When interface entry is deleted from interface-state OS,
in NAT InterfaceStateListener, as part of clean-up, we try to delete all
the SNAT sessions data assocaiated with that interface. But, an attempt was
done to delete an entry from snatint-ip-port-map even there doesn't exists any
SNAT translation resulting in this exception. Changes done to delete
snatint-ip-port-map ony if an valid entry exists.

Change-Id: Ic0d4fa08b91700c2b4a63ab6e613d7554741d413
Signed-off-by: cgowdru <chetan.arakere@altencalsoftlabs.com>