Remove unnecessary portOp error from subnetroute
This cannot find portOp Error was earlier relevant
where race conditions existed between interface-state
event southbound and subnet-port openstack configuraiton from
northbound.   Later we divided handling the subnetroute
creation into two batches to address the race condition
properly in [1].

As a result with [1] being merged, the error statement for
portOp is not required anymore and so this fix removes that
error statement.

[1] https://git.opendaylight.org/gerrit/#/c/61823/

Change-Id: Ic61f7ec5e3b9733f3e2f9a6bfc56af15c8af6d73
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>