Bug-9041: refactor addToNeutronRouterInterfacesMap
This optimise a code introduced in the commit
893706769d2a524c1b453f5473eb66d708497757 ("Bug 6920 : Fix for ACL
portSecurityUpdate to work with DjC + listed fixes"), in a part, lied to
write synchronisation for router-interfaces map object.

It seems to be better to get router interfaces list from
router-interfaces map object, if this object exists, update its port
list with a new interface and than update the object itself and rewrite
it in CONFIGURATIONAL DS.  Syncwrite/syncupdate operations only for new
added interface seems strange.

If router-interfaces map object still not exist, it will be created with
current router Id, updated with a new port list and than writed to
CONFIGURATIONAL DS as well.

Change-Id: I7e90d12ab1b01003e77493054794151da4ff2d57
Signed-off-by: Valentina Krasnobaeva <valentina.krasnobaeva@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>