Update patch set 1
Patch Set 1: Code-Review+1

These are learnt adjacencies. Hence. they are being added in addNewAdjToVpnInterface() method, which has an explicit if for PNF cases.
But, the deletes will have to be handled both during the removal of a leanrt PNF adjacency(1), and also during an external subnet delete(2).
(2) is what is being handled in this patch

Patch-set: 1