Bug 8860: NPE in getElanTag from SubnetmapChangeListener
The exception is genuine, but in order to enable tightening
of exception throws in netvirt, we are logging this
exception gracefully via this fix.

The real reason why this exception occurs can be ascertained
only the context information available when this issue happens
again in any upstream job.

The bug itself does not have link to logs that have this
exception thrown, for any postmortem/analysis.

Change-Id: I1c29efad5de85bfd340ea54a3600db9997c12a4b
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>
(cherry picked from commit 7c59b9ca35c1f40350bb7fed9747f380756de015)