vpnmanager: update vpn instance families based on subnetmap updates only
Upon a change in configuration, VPN instance from config DS is updated.
The change is due to a subnetmap moving to or from an other vpnid. If the old
VPN was a BGPVPN, then the list of subnetmaps will be parsed to check if
an IP family is missing or not. Reversely, the introduction of a new
subnet refreshes the IP families of the VPN instance.
According to the information, the VPN instance is updapted.
Consequently, if BGPVPN is configured, then the associated BGP VRFs are
updated accordingly.
Also, the vpnmanager shell command is modified accordingly with correct
AFConfig. Because a VPN instance can be either IPv4 or IPv6 or both,
then a check should be done against the good structure to parse.

Change-Id: I76b067eda44900d5c7bbf3f9792a55c32f711612
Signed-off-by: Valentina Krasnobaeva <valentina.krasnobaeva@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>