vpnservice: misc fixes mainly about BGPVPN configuration
- Bug-9134: avoid returning exception when checking for a non IPv4 prefix
In order to not having an exception returned, the Nwutils library is not
used to check for an IPv4 address.

- vpnmanager: for each vrf sub family deleted, partial FIB removed
Upon VRF sub family deletion, associate FIB entries are looked up and
removed.

- add generic vpn-config attribute in vpn-instance
This generic field contains BGPVPN information ( import / export rules,
RD), that will be used and synced with ipv4-family and ipv6-family in
the same structure. By default, vpn-config is present, and is used. If
an IPv4 subnet is found, then ipv4-family is created based on what
contains vpn-config.

- Bug 9168: multipath command supports vpnv6 address families
The multipath command changes:
o the addr-family is optional and takes as default value vpnv4
o it is possible to set multipath for vpnv6 , evpn, or lu address-family

- Bug-9167: in case of subnetmap update, the networkId is set correctly
There are cases where the subnetmap networkId field is flushed, and
never set back when an update occurs.

Change-Id: I3af7f6d5e700735450fd615d1b080229c2dfd771
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>