Update patch set 1
Patch Set 1:

Sam, thanks for this, but there would actually be A LOT more to back-port in controller (everything under https://git.opendaylight.org/gerrit/#/q/topic:bug/9060).  So you can merge this if you like, and you'll get Josh's original mdsaltrace, but not the trace:transactions stuff.  Backporting all of it is another few hours of work which is not a priority for me in the short-term.

Patch-set: 1