Update patch set 13
Patch Set 13:

(1 comment)

> (1 comment)
 > 
 > What is the case where br-int exists but without manager set? Are u
 > saying someone explicity went and removed this?

I assume you mean controller instead of manager. This is the way devstack used to work, it would create the bridge and set-manager but not set-controller. Most recently that function was put back in around 2 months ago. Overall, I guess the answer is that this is a legacy supported feature even though it doesn't quite make sense for an installer to work this way. Still, if we can support it, no reason not to.

Patch-set: 13