Update patch set 3
Patch Set 3:

> Curently in Controller based SNAT, if there exists the external-VPN
 > presence(FIP) on Non-Napt, and if there exists an valid vxlan
 > tunnel, then a remote FIB flow for SNAT IP will be programmed on
 > Non-NAPT and this new flow(21->group) is not required.
 > 
 > ex:
 > 
 > cookie=0x8000003, duration=4206.280s, table=21, n_packets=0,
 > n_bytes=0, priority=42,ip,metadata=0x30d46/0xfffffe,nw_dst=172.160.0.3
 > actions=set_field:0x11171->tun_id,load:0x300->NXM_NX_REG6[],resubmit(,220)
 > 
 > cookie=0x90186a5, duration=4991.870s, table=36, n_packets=6,
 > n_bytes=444, priority=5,tun_id=0x11171 actions=goto_table:44
Chetan,

In Conntrack based implementation for flat/vlan n/w we are not advertising the prefix to bgp manager and hence table 21 for reverse SNAT traffic flow will not be programmed by VrfEntryListener.
Do you think the  way it is addressed in this patch breaks any use case? We have to make the flat/vlan based implementation dependent on bgpManager and related classes otherwise. We do have separate patch to handle bgp use cases in a different class.

In controller snat I see that metadata("metadata=0x30d46/0xfffffe") is not of the external subnet and hence the reverse traffic for NFIP to FIP is not hitting this rule where the metadata will be external n/w after FIP translation, I am missing something here?

Patch-set: 3
Label: Verified=0