vpn-instance: add generic vpn-config attribute in vpn-instance
This generic field contains BGPVPN information ( import / export rules,
RD), that will be used and synced with ipv4-family and ipv6-family in
the same structure. By default, vpn-config is present, and is used. If
an IPv4 subnet is found, then ipv4-family is created based on what
contains vpn-config.

Change-Id: I08173aa6523b6d5d73e7e3b3288b26e64529902a
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>