Update patch set 1
Patch Set 1:

> I think we should avoid adding “throws Exception” declarations in
 > interfaces — let’s make them more specific... 

I actually disagree with this I think; in a call stack who's only purpose it is ultimately get invoked by e.g. a doExecute() throws Exception, to me "throws Exception" is just fine - just like in a @Test or a main().  

But perhaps we can agree to disagree on that detail and move forward with wrapping this one up anyway - I've adapted it now - OK for you like this?

 > (It’s also worth thinking about the InterruptedException handling.)

Later? ;)

Patch-set: 1
Label: Verified=0