Update patch set 1
Patch Set 1:

> Shouldn't we look at this problem from a more holistic perspective
 > across all implications instead of just L2?

Vivek, do you mean coordinating flowrefs across all the modules? If so, I agree that needs to be fixed. Started a thread with Vishal and you on a spec to do that work. I think we can make a quick patch here to fix the bug and work on the spec as another patch.

Patch-set: 1