Update patch set 118
Patch Set 118:

Hi Sam,

There are couple more same exceptions that comes, when the interface is added to the router.

However, we will look into this and fix it as part of 9181 and 9182.

We would need to merge this in to make forward progress for other review patches in IPv6.

Thanks in advance for your help.

Patch-set: 118