WIP: This flow is not present on OVS after “upgrade”:
WIP: DO NOT COMMIT!!!

table=21, priority=10,ip,metadata=0x30d42/0xfffffe actions=group:225000
(checks the VRF id and forwards to a group
that outputs to the external interface)
The flow is in fact written by ODL but rejected by OVS because
group:225000 is not present at the time the flow is written.
Group:225000 is not written because at the time that is triggered
(twice, actually) there is no /elan-interfaces/elan-interface for the
external network (trunk) port because OVS has not yet connected.

Change-Id: Id45e57cdc85c9b53094fce9163094f2711b0ab76
Signed-off-by: Josh <jhershbe@redhat.com>