vpnmanager: route interface state applies to more than one IP.
The introduction of dual stack makes it possible to handle more than one
fixed IP, then more than one associated subnet. This change applied the
change of yang port-op-data-entry structure to support a list of
subnet-ids. This information is updated in
SubnetRouteInterfaceStateChangeListener, where the list of subnets is
extracted from the Neutron port.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Change-Id: I2936014f9aaa5b42b820466affd3ada31695860d