vpnmanager: misc fixes related to dual stack per neutron port.
- route interface state applies to more than one IP.
introduction of dual stack makes it possible to handle more than one
fixed IP, then more than one associated subnet. This change applied the
change of yang port-op-data-entry structure to support a list of
subnet-ids. This information is updated in
SubnetRouteInterfaceStateChangeListener, where the list of subnets is
extracted from the Neutron port.

- DHCP secondary IP adjacency has no nextHopIP provisioned
This commit fixes the case where a secondary subnetwork IPv4 or IPv6 is
appended on the same network. In that case, the adjacency provisioned
should inherit from the same nextHop IP address as the first adjacency.
This calculation is done like it is done at the creation of VPN
adjacencies, when the first DHCP port is created.

- Bug-9274 : exception optimisticlockfailed on deleting "unbound vpn
  service from interface"

Change-Id: I2936014f9aaa5b42b820466affd3ada31695860d
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>