vpnmanager: misc fixes related to dual stack per neutron port.
- route interface state applies to more than one IP.
introduction of dual stack makes it possible to handle more than one
fixed IP, then more than one associated subnet. This change applied the
change of yang port-op-data-entry structure to support a list of
subnet-ids. This information is updated in
SubnetRouteInterfaceStateChangeListener, where the list of subnets is
extracted from the Neutron port.
- vpnmanager: remove and update prefixtointerface for dual stack addresses
Due to the possibility to have two primary IP addresses per port, some
VPN migration scenarios may not work with the other IP address. The code
is adapted to parse all the primary adjacencies of the VPN Interface.

Change-Id: I2936014f9aaa5b42b820466affd3ada31695860d
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>