vpnmanager: DHCP secondary IP adjacency has no nextHopIP provisioned
This commit fixes the case where a secondary subnetwork IPv4 or IPv6 is
appended on the same network. In that case, the adjacency provisioned
should inherit from the same nextHop IP address as the first adjacency.
This calculation is done like it is done at the creation of VPN
adjacencies, when the first DHCP port is created.

Change-Id: Ifeb068d6588b1d5a1c2b627566323016237a3bbb
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>