Bug 8964 - Neutron test neutron.tests.tempest.scenario.test_floatingip.FloatingIpSameNetwork.test_east_west fails
Set the in_port to zero in NAPT_PFIB table. This prevents the packet
drop in NAPT switch when the packet originating from tunnel is send back
to the same tunnel after DNAT to a Non Napt Switch.

The test[1](FloatingIpSameNetwork) verifies the traffic between a vm
with NFIP and vm with FIP with same internal n/w. The test now fails if
the NFIP vm and FIP vm happens to be in the same non NAPT switch.

[1]https://github.com/openstack/neutron/blob/master/neutron/tests/tempest/scenario/test_floatingip.py#L125 

Change-Id: I42d98cc411ae21b8d32e44f4bc9217e22946cdc8
Signed-off-by: Aswin Suryanarayanan <asuryana@redhat.com>