BUG 9220: don't use tun_gpe_np as match field
tun_gpe_np is not supported as match field on nsh patch OVS 2.6. This
field was used to match nsh vxlan encapsulated packets on sfc classifier
pipeline. Even if tun_gpe_np could be used, it wouldn't help as
currently encapsulation is implemented as vxlan+eth+nsh as opposed to
vxlan+nsh.

Workaround is to check for eth+nsh and check for tun_dst=0 with higher
priority to discern non tunneled nsh packets against tunneled ones.

Change-Id: I5038f5d8e89bdada559a736c711478f1bf123453
Signed-off-by: Jaime Caamaño Ruiz <jcaamano@suse.com>