[WIP] ping6 support for router interface on a _multi_ network
Upon some investigation it was seen that fibentry:vrfTables were not
updated with the second Ipv6 subnet that is added to the neutron router.
Also, the odl-l3vpn:adjacency information for the router-interface is
not being updated. Because of this, ping6 (auto-responder) flows were
not programmed in FibTable(i.e., Table21). This patch addresses these
issues and I was able to validate that ping6 is working for the second
IPv6 subnet aswell. However, there is a possiblity that this patch could
lead to some regressions. I would appreciate some review comments from
people who are more familiar with the code that is being updated.

Change-Id: I5bff26b1fee616d37731a951b612fc310730d9a6
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>