Convert vpnmanager-impl to use mdsal EOS APIs
The controller EOS APIs are deprecated.

The ArpMonitoringHandler previously used the EntityOwnerUtils to
register its candidate which has the side effect of registering
a global EntityOwnershipListener that maintains local ownership
state for all entities in a static cache. It also stores change
events in a static (unsycnhronized) EVENTS_HISTORY list which grows
unbounded. It seems this is only used in one place to print the
history via a CLI command. Both these side effects are problematic
and I don't think need/should be propagated forward.

Change-Id: Ied8e3f061c798de9a0e8af97e13fc9dd7e5d51a6
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>