Update patch set 1
Patch Set 1:

> some rework is up to be pushed on master from https://git.opendaylight.org/gerrit/#/c/57980
 > this fix should be enqueued.

I can rebase on top of that patch. The reason I got into this code is that I need to change it to use the new MDSAL EntityOwnershipService APIs as the controller APIs are deprecated and scheduled for removal. However the static nature of this class is problematic for that work.

Patch-set: 1