Ignore addElanInterface
Autorelease is failing on this test with
org.opendaylight.yangtools.testutils.mockito.UnstubbedMethodException: getInterfaceInfoFromOperationalDataStore(java.lang.String interfaceName, org.opendaylight.genius.interfacemanager.globals.InterfaceInfo$InterfaceType interfaceType) is not implemented in testInterfaceManager

Once this method is included in testinterfaceManager then all the subsequent methods start to fail since they are null or not implemented either.

This patch below reworks all this so the test will be reenabled at that time.

https://git.opendaylight.org/gerrit/62894

Change-Id: Ia1166d4055804733adf3fc6e8a0aa376a520e41a
Signed-off-by: Sam Hague <shague@redhat.com>