vpnmanager: add ipv4 and ipv6 sub-services for L3VPN
In the case with two router solution, two default VPNids have to be
chosen, so that outgoing packets can go to the corret VPN. Indeed, the
solution is chosen to select the VPN IPv6 ( or arbitrarily choose one)
with the highest priority, on table 17. Then packets reach table 19. If
there are no matches in table 19, then packet will be resubmitted to
table 19, where a second rule with a second VPN (IPv4) will be set.

Change-Id: I9981db0308402099901671533eeb25dead26daa8
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>