netvirt-964 : upon detach IPv4 router port, IPv6 entries are removed too
This is related to a dual stack scenario. When some ports are configured
with dual stack IPs, then if someone removed the IPv4 router port, then
a check must be done when removing the interfaces related to the subnet
of that IPv4 router port. Indeed, if the port is shared with other
subnets, then it must not be removed.

Change-Id: Iea5f574b6ebe9a1401bda98fa5bb9949c68c3cf1
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>