Update patch set 36
Patch Set 36:

(13 comments)

I hope I answered to most of your remarks.

I will have 2 if() functions that will respectively check for addition or deletion.
This is necessary because both functions do not do the same work.

I would like to clarify what is really needed for this review, from what is relying on some other design changes that have not been taken into account in none of the reviews.

for example, getVpnRd() should be part of a separate review.
It is not directly related to the review.

also, the getSubnetmap reduction should be a separate review too.

And I consider both reviews are low priority ( unless it brings exceptions).

thanks for pushing a new review.

Patch-set: 36