Fix FindBugs violations and enable enforcement
The thrift-generated code has a quite a few violations and, as
we do with other generated code, I excluded it from analysis.
Only one of the violations in was serious:

  Impossible cast from org.apache.thrift.TApplicationException to
  org.apache.thrift.TBase

This occurs in every BgpConfigurator inner class that creates an
anonymous AsyncMethodCallback instance from its getResultHandler
method. The generated code in onError method does a blind cast
to a clearly incompatible class. I'm not sure what the correct
solution is if we wanted to manual fix the generated code as I'm
not familiar with how thrift works. There isn't a TBase
implementation for an Exception and the generated *_result classes
don't have a field for an Exception.

Change-Id: I3a6a2fd60f2088e307a213563f267bdf00bdd1ef
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>