21->nat-group flow is not present on OVS after “upgrade”
table=21, priority=10,ip,metadata=0x30d42/0xfffffe actions=group:225000
(checks the VRF id and forwards to a group
that outputs to the external interface)
The flow is in fact written by ODL but rejected by OVS because
group:225000 is not present at the time the flow is written.
Group:225000 is not written because at the time that is triggered
(twice, actually) there is no /elan-interfaces/elan-interface for the
external network (trunk) port because OVS has not yet connected.

Create a default NAT group(225000) before installing flow the
21->group:225000. If elan-interface(for external-interfaces) are missing
during this group creation, then create this group with an drop action.
Once elan-interfaces are available, this group will be updated with proper
action values.

Steps perfomed
--------------

1) Shutdown ODL (logout)
2) wipe data store (rm -rf $KARAF_HOME/journal/* &
	$KARAF_HOME/snapshots/*)
3) disconnect OVS and clear groups/flows
	sudo ovs-ofctl del-flows -O Openflow13 br-int
	sudo ovs-ofctl del-groups -O Openflow13 br-int
	sudo ovs-vsctl del-manager
	sudo ovs-vsctl del-controller br-int
4) restart ODL (./karaf)
5) Wait for net-odl full sync from neutron
	- waited until DS gets poulatedi
	- (ex : GET http://192.168.56.1:8181/restconf/config/neutron:neutron/ports)
6) reconnect OVS
	sudo ovs-vsctl set-manager "tcp:192.168.56.1:6640"
	sudo ovs-vsctl set-controller br-int "tcp:192.168.56.1:6653"

Change-Id: Iad89b3ff7935846b0bf1a3b9b3a4690ff87fcd9c
Signed-off-by: cgowdru <chetan.arakere@altencalsoftlabs.com>