Bug 9297 : Adjacency doesn't exist exception fo FIP
Description : When are ARP request for floating-ip(which is learnt
earlier) is received(with mac address change due to deletion and creation
of same FIP during tempest), query is done to delete the odl-adjacency
from vpn-interface which doesn't exists.

Instead of quering all adjacenies, change done for existance of specific
adjancency and delete it if only exists.

Change-Id: Id39f9f0830330bb84869c0d3a88c4787cd55df3c
Signed-off-by: cgowdru <chetan.arakere@altencalsoftlabs.com>