Update patch set 4
Patch Set 4:

(1 comment)

Has any testing been done with the gate jobs to ensure that the existing csit tests do not break with this change?

In the comments below is a detail about using the multi-patch - does this mean there are dependency patches? If so please add a Depends-on: xxx to the commit-message. And make sure to run the csit tests with a comment added to the patch: "test- netvirt- current" - without the quotes and spaces removed. If that keyword is seen anywhere in a comment it will trigger the csit.

Patch-set: 4