Convert vpnmanager-impl to use JobCoordinator
I changed the classes using JobCoordinator to use the injection
annotations and removed them from the blueprint XML. However that
changed the bean names so other beans failed. So I ended up changing
all the beans to use the annotations which is what we want anyway to
make things easier going forward anyway.

Change-Id: I9b1753e88c24ed5a533d925dc940fcfee04fc13f
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>