Remove NeutronUtils#lock/unLock in favor of KeyedLock
The lock/unLock code in NeutronUtils should not be static and
doesn't really belong in a static utils class. A new KeyedLock
class was aded to infrautils so utilize that in NeutronvpnManager
and NeutronFloatingToFixedIpMappingChangeListener. Instead of a
shared, global KeyedLock, each class creates local instances
per resource to lock (router, vpn, interface).

Change-Id: Icbd9c341c8d7399e31ed8c38b92177a7fa8ed666
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>