Eliminate statistics-impl dependency on vpnmanagr-impl
statistics-impl depends on vpnmanagr-impl to get access to the
InterfaceUtils class. This is problematic for converting vpnmanagr-impl
to use blueprint injection annotations b/c all the @Singleton classes
get included in the statistics-impl blueprint xml. It appears the
plugin scans @Singleton classes in all dependencies. I don't know if
there's a way to prevent that but eliminating the dependency on
vpnmanagr-impl would fix it. It is better not to depend on an impl
bundle anyway. Therefore I moved the InterfaceUtils class to the
vpnmanagr-api bunndle. In addition I modified the maven-bundle-plugin
instructions to not export any packages to prevent a dependency on
vpnmanagr-impl in the future.

Change-Id: I3731a4faf37a1ba56659346cb0a5697707acbc32
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>