Update patch set 99
Patch Set 99:

59413 review, along with parent dependencies:
- 59445
- 59700
- 57780
need to be merged together because of dependencies.

Indeed, a choice had to be done.
a- either the changes were done in a single revision
b- or the changes are split in several parts

b) solution was chosen, and some issues can be seen, but are temporary.
for instance 59700 has issues that only 59445 can fix ( for example, writing/reading a vpn interface , whereas a vpn interface op entry is present, is a transitional state.

Patch-set: 99
Reviewer: Gerrit User 6574 <6574@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0