Start to clean up ElanInstanceManager sync/async mess
This change does (should..) not actually change any behavior; what is de
facto in the current code already sync is kept sync, and some IMHO
pointless async DJC usages are removed.  Retries are preserved.

Background: ElanUtils' waitForTransactionToComplete() is synchronously
waiting anyway (note that it does a get()) it IMHO makes absolutely no
sense to return an already completed Future from it, and then pass than
back through up the call chain and pretend it's async - it's clearly
not.

The ripple effect of this is that we find several usages of the
JobCoordinator which IMHO are totally pointless.  If the only reason for
those is the retries more than the asynchronicity, which at least in the
particular cases found here re pointless anyway due to above, then IMHO
instead of the DJC (JC) it is better and simpler to just use the new
RetryingManagedNewTransactionRunner instead, which this does in a few
places.  (But there is more work to be done around this topic all over
the place.)

This came out of and is motivated by findings in
https://jira.opendaylight.org/browse/NETVIRT-937 and the comment from
Tom in https://git.opendaylight.org/gerrit/#/c/65821/ and is related to
https://lists.opendaylight.org/pipermail/netvirt-dev/2017-November/005829.html

Change-Id: I22e16f90cf4cc1f13aa7dce7e80a125a16f62c9c
Signed-off-by: Michael Vorburger <vorburger@redhat.com>