Fix FindBugs violations and enable enforcement in ipv6service
See in-line comments.

Eliminated the static instances for IfMgr and Ipv6PeriodicTrQueue.

The ShowIpv6Command class referenced the static IfMgr instance.
I extracted an interface, ElementCache, for the IfMgr whicb is
advertsed as a service and injected into ShowIpv6Command. The shell
bundle shouldn't reference packages in the impl bundle so I also
extracted interfaces for VirtualRouter, VirtualPort et al.

Change-Id: I635ff43cffded41c7a32079214cff48a057f4fac
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>