Convert AclInterfaceCacheUtil to non-static singletoni service.
An AclInterfaceCache interface was added to the api bundle with
the AclInterfaceCacheImpl in the impl bundle. The AclInterfaceCacheImpl
is advertised as an OSGi service.

The AclInterface class was not thread-safe. Instead of making it thread-safe,
it was made immutable to simplify things. The cache creates new instances
on updates. Also AclInterface instances are now created via a Builder.

Change-Id: Id3b69a87795d6ff0c6dcbd3ac599cd0326077e25
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>