Update patch set 2
Patch Set 2: Code-Review-1

Looking at the call sites, it's normal for underlayNetworks to be empty. So checking for empty underlayNetworks is a convenience for callers otherwise all sites would have to check to avoid an error logged. I think we should leave it as is. As for null, we could require @Nonnull or make it @Nullable.

Patch-set: 2
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1