Update patch set 2
Patch Set 2:

(1 comment)

Tom, LGTM, but just to +1 this in full confidence: Re. the CLI commands of this bundle, I've no idea if CSIT exercises them.  You wouldn't want to build a distro with this and just do a 5' manual verification of the CLI commands and let us know here just that they are still at least available, would you?  (PS: I've long wanted to add a util to itestutils to make writing IT for CLI commands easy; let me know if you have nothing else to do and want to help with that, some time.)

Patch-set: 2