Update patch set 2
Patch Set 2:

> > @tom: did you get a chance to verify the cli's still worked?
 > 
 > Not yet. They def wire correctly. Not sure I can really test
 > functionality b/c I really don't know what most do.
 > 
 > > https://git.opendaylight.org/gerrit/66147 this patch removed
 > > annotations to get the cli working.
 > 
 > that patch looks unrelated

Ok, so wiring looked good, means this started fine in karaf and could be called? That's all we need.

The other comment about the patch, was just that in the patch the annotations were removed in favor of blueprint xml to get the cli working, so just wanted some more eyes on that patch.

Patch-set: 2