Update patch set 9
Patch Set 9:

@tom, @michael, I think there is something off in this patch. Thsi one fails to get all vm ips for the first network in both the l2 and l3 test suites. So it does seem like maybe an initial event is not happening, but eventually does by the time the vms are started on the second network.

Maybe look at the tunnel files, anything with tep, tunnel or zone. Tunnels are built initially as vms are spawned so maybe the tunnel took a while to create. Other things around the first network, is that it is a vlan network rather than vxlan.

Patch-set: 9
Label: Verified=0