Update patch set 12
Patch Set 12:

> Sam, are we planning to get this in soon? If this will take more
 > time, shall i push another patch with pom.xml change alone, because
 > diagstatus is currently missing ELAN in output because annotations
 > are not detected. :) We need it asap, for some of our testing

We haven't gotten CSIT to pass and it's still unclear if it's this patch or infra. As it stands now I have reverted back to using <bean> elements but CSIT run was red (looks like infra issue). 

I'd rather get this patch working so if you or someone on your team that knows this code could help that would be great.

Patch-set: 12
Label: Verified=0