Update patch set 12
Patch Set 12:

wow it passed CSIT. So there seems to some issue with the annotations unless it's just a coincidence it passed. Assuming it's the former, I could use help from folks who know the code otherwise I'm stabbing in the dark. So based on which tests fail, what part of elanmanager code could cause it?  It seems either something isn't getting instantiated or started - but what?

Patch-set: 12