Convert L2GatewayCacheUtils to non-static singleton service
Crrated an interface, L2GatewayCache, with an associated
L2GatewayCacheImpl that is advertised as an OSGI service.
The L2GatewayCacheUtils had a couple update* methods that
only had one call site so I just inlined the code in the
caller to keep L2GatewayCache as a simple cache.

To avoid perpetuating circular bundle dependencies and having to
use optional availability on reference import, I put the
L2GatewayCacheImpl in a new bundle, cache-impl.

Change-Id: I4c1dcc761afcdaa214e4bdcf53f9ddb85363539f
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>