Upgrade: race cond. btwn elan-dpn-interfaces and arp-responder for external ifaces
Depends on this gerrit from genius:
https://git.opendaylight.org/gerrit/#/c/65894/

A comment from inside the code:
The following through the end of the function deals with
an upgrade scenario where the neutron configuration
is restored before the OVS switches reconnect. In such a
case, the elan-dpn-interfaces entries will be
missing from the operational data store. In order to
mitigate this we use DataTreeEventCallbackRegistrar
to wait for the exact operational md-sal object we need
to contain the external interface we need.

Change-Id: Id02564a13e7f85933867e21c51e36c32bc60cd4e
Signed-off-by: Josh <jhershbe@redhat.com>