Bug 8998 - Stale l2gw connection cleanup
upon cluster reboot if the device is not connected
it is not shown up in the cache or oper datastore.

Now when l2gw connection delete request is received it has no clue
on which device it has to delete the things and ignores the
delete request.

when the device connects later, cleanup job is scheduled
which will look for stale logial switches and deletes them.

any logical switch which is not present in networks dump or
l2gwconnections dump is treated as stale logical switch.

cleanup job is scheduled instead of firing immediately because
the network datastore may not be loaded yet as the cluster may
be just started.

cleanup if the device never shows up is not done.

Change-Id: I84d5e956ab6a94988f9ddf8faa8f3cf62f5c07d1
Signed-off-by: eaksahu <a.k.sahu@ericsson.com>