Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> (1 comment)

The optional is probably best at this point since we could gt it done. A major reorg might be difficult at this point. We need to get the dualstack stuff in and sounds like a reorg would delay that further.

Michael went over the cache utils last week - did that not go over well or not apply here?

Patch-set: 2
Label: Verified=0